requirements.txt
streamlit
transformers
torch
fer
opencv-python
pillow
librosa
soundfile
numpy
pandas
plotly